+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 2024. 11. 01
 - Added `Item` model with `migrations`
 - Added item `services` and `repositories`
+- Added `ItemController`, `routes` and `requests`
+- Bind new `services` and `repositories`
 
 ## 2024. 10. 24
 ### @tbence